СЂР°Р·РѕР±СЂР°РЅРЅР°РєРѕРЅРµС†  РєРѕРµ-РєР°Рє РєР°Р»СЊРєСѓР»СЏС‚РѕСЂ СЃ Р°РґР°РїС‚Р°СЂРѕРј. РЅР°РґРѕ РµС‰Рµ РґРѕСЂР°Р·Р±РёСЂР°С‚СЊСЃСЏ СЃ РѕС‡РµСЂРµРґРЅРѕСЃС‚СЊСЋ

+++ home_work_3/calcs/additional/CalculatorWhithCounterClassic.java
@@ -30,7 +30,7 @@ public class CalculatorWhithCounterClassic extends CalculatorWithMathExtends {
     }
 
     @Override
-    public void setFD(double f, int d) {
+    public void setFD(double f, double d) {
         super.setFD(f, d);
     }
