+++ home_work_3/calcs/additional/CalculatorWithCounterAutoAgregationDecorator.java
@@ -38,7 +38,7 @@ public class CalculatorWithCounterAutoAgregationDecorator implements ICalculator
         return this.icalculator.getDel();
     }
     @Override
-    public void setFD(double f, int d) {
+    public void setFD(double f, double d) {
 
         this.icalculator.setFD(f, d);
