+++ home_work_3/calcs/additional/CalculatorWithCounterAutoSuper.java
@@ -40,7 +40,7 @@ public class CalculatorWithCounterAutoSuper extends CalculatorWithMathExtends {
     }
 
     @Override
-    public void setFD(double f, int d) {
+    public void setFD(double f, double d) {
         super.setFD(f, d);
     }
